refactor: separate updateEvents & use droplist for event options
