+++ src/routes/console/project-[project]/functions/function-[function]/settings/updateEvents.svelte
@@ -1,0 +1,144 @@
+<script lang="ts">
+    import CardGrid from '$lib/components/cardGrid.svelte';
+    import Empty from '$lib/components/empty.svelte';
+    import Heading from '$lib/components/heading.svelte';
+
+    import { invalidate } from '$app/navigation';
+    import { page } from '$app/stores';
+    import { trackEvent } from '$lib/actions/analytics';
+    import { Dependencies } from '$lib/constants';
+    import Form from '$lib/elements/forms/form.svelte';
+    import { TableCell, TableCellText } from '$lib/elements/table';
+    import TableList from '$lib/elements/table/tableList.svelte';
+    import { symmetricDifference } from '$lib/helpers/array';
+    import { addNotification } from '$lib/stores/notifications';
+    import { sdkForProject } from '$lib/stores/sdk';
+    import { writable, type Writable } from 'svelte/store';
+    import { func } from '../store';
+    import { EventModal } from '$lib/components';
+    import { Button } from '$lib/elements/forms';
+    import DropList from '$lib/components/dropList.svelte';
+    import DropListItem from '$lib/components/dropListItem.svelte';
+
+    const functionId = $page.params.function;
+    const eventSet: Writable<Set<string>> = writable(new Set($func.events));
+    let showEvents = false;
+    let eventValue: string;
+    let isDisabled = true;
+    let showDropdown: boolean[] = [];
+
+    async function updateEvents() {
+        try {
+            await sdkForProject.functions.update(
+                functionId,
+                $func.name,
+                $func.execute,
+                Array.from($eventSet),
+                $func.schedule,
+                $func.timeout,
+                $func.enabled
+            );
+            invalidate(Dependencies.FUNCTION);
+            addNotification({
+                message: 'Permissions have been updated',
+                type: 'success'
+            });
+            trackEvent('submit_function_update_events');
+        } catch (error) {
+            addNotification({
+                message: error.message,
+                type: 'error'
+            });
+        }
+    }
+
+    function handleEvent(event: CustomEvent) {
+        if (eventValue) {
+            const newSet = new Set(
+                [...$eventSet].map((e) => (e === eventValue ? event.detail : e))
+            );
+            eventSet.set(newSet);
+        } else {
+            eventSet.set($eventSet.add(event.detail));
+        }
+        eventValue = null;
+    }
+
+    $: if ($eventSet) {
+        if (symmetricDifference(Array.from($eventSet), $func.events).length) {
+            isDisabled = false;
+        } else isDisabled = true;
+    }
+
+    $: if (!showEvents) {
+        eventValue = undefined;
+    }
+</script>
+
+<Form on:submit={updateEvents}>
+    <CardGrid>
+        <Heading tag="h6" size="7">Update Events</Heading>
+        <p>Set the events that will trigger your function. Maximum 100 events allowed.</p>
+        <svelte:fragment slot="aside">
+            {#if $eventSet.size}
+                <TableList>
+                    {#each Array.from($eventSet) as event, i}
+                        <li class="table-row">
+                            <TableCellText title="id">
+                                {event}
+                            </TableCellText>
+                            <TableCell showOverflow title="options" width={40}>
+                                <DropList
+                                    bind:show={showDropdown[i]}
+                                    placement="bottom-start"
+                                    noArrow>
+                                    <button
+                                        class="button is-text is-only-icon"
+                                        aria-label="more options"
+                                        on:click|preventDefault={() =>
+                                            (showDropdown[i] = !showDropdown[i])}>
+                                        <span class="icon-dots-horizontal" aria-hidden="true" />
+                                    </button>
+                                    <svelte:fragment slot="list">
+                                        <DropListItem
+                                            icon="pencil"
+                                            on:click={() => {
+                                                showDropdown[i] = false;
+                                                showEvents = true;
+                                                eventValue = event;
+                                            }}>
+                                            Edit
+                                        </DropListItem>
+                                        <DropListItem
+                                            icon="trash"
+                                            on:click={async () => {
+                                                $eventSet.delete(event);
+                                                eventSet.set($eventSet);
+                                            }}>
+                                            Delete
+                                        </DropListItem>
+                                    </svelte:fragment>
+                                </DropList>
+                            </TableCell>
+                        </li>
+                    {/each}
+                </TableList>
+
+                <div class="u-flex u-margin-block-start-16">
+                    <Button text noMargin on:click={() => (showEvents = true)}>
+                        <span class="icon-plus" aria-hidden="true" />
+                        <span class="u-text">Add event</span>
+                    </Button>
+                </div>
+            {:else}
+                <Empty on:click={() => (showEvents = true)}>Add an event to get started</Empty>
+            {/if}
+        </svelte:fragment>
+
+        <svelte:fragment slot="actions">
+            <Button disabled={isDisabled} submit>Update</Button>
+        </svelte:fragment>
+    </CardGrid>
+</Form>
+
+<EventModal bind:show={showEvents} initialValue={eventValue} on:created={handleEvent} />
